Remove form_controller
There are two form_controller classes which inherit from each other in
surprising ways.

This commit:

- Remove the unused routes in form controller
- Moves the useful routes in form controller into forms/form_controller
- Rename forms/form_controller forms/base_controller
- Make sure all controllers inherit from base_controller
