+++ app/controllers/forms/privacy_page_controller.rb
@@ -1,5 +1,5 @@
 module Forms
-  class PrivacyPageController < FormController
+  class PrivacyPageController < BaseController
     def show
       @privacy_policy_url = current_context.privacy_policy_url
     end
